+++ src/Symfony/Component/Workflow/MarkingStore/MarkingStoreInterface.php
@@ -36,8 +36,7 @@ interface MarkingStoreInterface
     /**
      * Sets a Marking to a subject.
      *
-     * @param object  $subject A subject
-     * @param Marking $marking A marking
+     * @param object $subject A subject
      */
-    public function setMarking($subject, Marking $marking);
+    public function setMarking($subject, Marking $marking, array $context = []);
 }