+++ src/Symfony/Component/Workflow/Tests/MarkingStore/MethodMarkingStoreTest.php
@@ -1,0 +1,74 @@
+<?php
+
+namespace Symfony\Component\Workflow\Tests\MarkingStore;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Workflow\Marking;
+use Symfony\Component\Workflow\MarkingStore\MethodMarkingStore;
+
+class MethodMarkingStoreTest extends TestCase
+{
+    public function testGetSetMarkingWithMultipleState()
+    {
+        $subject = new Subject();
+
+        $markingStore = new MethodMarkingStore(false);
+
+        $marking = $markingStore->getMarking($subject);
+
+        $this->assertInstanceOf(Marking::class, $marking);
+        $this->assertCount(0, $marking->getPlaces());
+
+        $marking->mark('first_place');
+
+        $markingStore->setMarking($subject, $marking);
+
+        $this->assertSame(['first_place' => 1], $subject->getMarking());
+
+        $marking2 = $markingStore->getMarking($subject);
+
+        $this->assertEquals($marking, $marking2);
+    }
+
+    public function testGetSetMarkingWithSingleState()
+    {
+        $subject = new Subject();
+
+        $markingStore = new MethodMarkingStore(true);
+
+        $marking = $markingStore->getMarking($subject);
+
+        $this->assertInstanceOf(Marking::class, $marking);
+        $this->assertCount(0, $marking->getPlaces());
+
+        $marking->mark('first_place');
+
+        $markingStore->setMarking($subject, $marking);
+
+        $this->assertSame('first_place', $subject->getMarking());
+
+        $marking2 = $markingStore->getMarking($subject);
+
+        $this->assertEquals($marking, $marking2);
+    }
+}
+
+final class Subject
+{
+    private $marking;
+
+    public function __construct($marking = null)
+    {
+        $this->marking = $marking;
+    }
+
+    public function getMarking()
+    {
+        return $this->marking;
+    }
+
+    public function setMarking($marking)
+    {
+        $this->marking = $marking;
+    }
+}