+++ src/Symfony/Component/Workflow/WorkflowInterface.php
@@ -58,7 +58,7 @@ interface WorkflowInterface
      *
      * @throws LogicException If the transition is not applicable
      */
-    public function apply($subject, $transitionName);
+    public function apply($subject, $transitionName, array $context = []);
 
     /**
      * Returns all enabled transitions.